// apps/backend/src/common/security-events/enums/security-event.enum.ts
export enum SecurityEventType {
  AUTH_REGISTER_SUCCEEDED = 'AUTH_REGISTER_SUCCEEDED',
  AUTH_REGISTER_FAILED = 'AUTH_REGISTER_FAILED',
  AUTH_LOGIN_SUCCEEDED = 'AUTH_LOGIN_SUCCEEDED',
  AUTH_LOGIN_FAILED = 'AUTH_LOGIN_FAILED',
  AUTH_LOGIN_RATE_LIMITED = 'AUTH_LOGIN_RATE_LIMITED',
  AUTH_REFRESH_SUCCEEDED = 'AUTH_REFRESH_SUCCEEDED',
  AUTH_REFRESH_FAILED = 'AUTH_REFRESH_FAILED',
  AUTH_REFRESH_REUSE_DETECTED = 'AUTH_REFRESH_REUSE_DETECTED',
  AUTH_REFRESH_RATE_LIMITED = 'AUTH_REFRESH_RATE_LIMITED',
  AUTH_LOGOUT_SUCCEEDED = 'AUTH_LOGOUT_SUCCEEDED',
  AUTH_LOGOUT_FAILED = 'AUTH_LOGOUT_FAILED',
  AUTH_PASSWORD_RESET_REQUESTED = 'AUTH_PASSWORD_RESET_REQUESTED',
  AUTH_PASSWORD_RESET_SUCCEEDED = 'AUTH_PASSWORD_RESET_SUCCEEDED',
  AUTH_PASSWORD_RESET_FAILED = 'AUTH_PASSWORD_RESET_FAILED',
  AUTH_PASSWORD_RESET_RATE_LIMITED = 'AUTH_PASSWORD_RESET_RATE_LIMITED',
  AUTH_PASSWORD_CHANGED = 'AUTH_PASSWORD_CHANGED',
  AUTH_PASSWORD_CHANGE_FAILED = 'AUTH_PASSWORD_CHANGE_FAILED',
  USER_ACCOUNT_LOCKED = 'USER_ACCOUNT_LOCKED',
  USER_ACCOUNT_UNLOCKED = 'USER_ACCOUNT_UNLOCKED',
  USER_ROLE_CHANGED = 'USER_ROLE_CHANGED',
  AUTH_ORIGIN_REJECTED = 'AUTH_ORIGIN_REJECTED',
  AUTH_RATE_LIMIT_EXCEEDED = 'AUTH_RATE_LIMIT_EXCEEDED',
  UPLOAD_RATE_LIMIT_EXCEEDED = 'UPLOAD_RATE_LIMIT_EXCEEDED',
  RATE_LIMIT_EXCEEDED = 'RATE_LIMIT_EXCEEDED',
}